refactor(ui): remove redundant data-wallet-ui attributes from components
The `data-wallet-ui` attribute is meant to be added by library users to a parent
element that contains use-wallet-ui components in non-Tailwind apps. The
attribute was incorrectly added to individual components initially, but the CSS is
defined using child selectors (`[data-wallet-ui] .class-name`), making these
component-level attributes redundant.

The attribute is now only used in modal components that use portals, where it
cannot be applied by the library user's parent element.